refactor: Sort users by last message time in ChatsFragment
This commit refactors the ChatsFragment class to sort the users by their last message time. The method sortUsersByLastMsgTime1() is added to sort the allUsers list in descending order based on the lastMsgTime property of each user. This ensures that the users are displayed in the correct order in the UI.