```text chore: Add CountryCodePicker library for phone number input
This commit adds the CountryCodePicker library (version 2.7.3) to the project's dependencies in the build.gradle file. The library provides a convenient way to select and display country codes for phone number input. This will be used in the PhoneNumberActivity to allow users to enter their phone numbers with the correct country code.